refactor: organize hooks with index.ts
