+++ pokedex/src/hooks/pokemon/index.ts
@@ -1,0 +1,7 @@
+export { usePokemonQuery } from "./usePokemonQuery";
+export { usePokemonEvolutionChainQuery } from "./usePokemonEvolutionChainQuery";
+export { usePokemonSpeciesQuery } from "./usePokemonSpeciesQuery";
+export { usePokemonWeaknessesQuery } from "./usePokemonWeaknessesQuery";
+export { usePokemonsQuery } from "./usePokemonsQuery";
+export { usePokemonsPageQuery } from "./usePokemonsPageQuery";
+export { usePokemonsByTypeQuery } from "./usePokemonsByTypeQuery";
