--- pokedex/src/hooks/usePokemon.ts
@@ -1,7 +1,0 @@
-export { usePokemonQuery } from "./pokemon/usePokemonQuery";
-export { usePokemonEvolutionChainQuery } from "./pokemon/usePokemonEvolutionChainQuery";
-export { usePokemonSpeciesQuery } from "./pokemon/usePokemonSpeciesQuery";
-export { usePokemonWeaknessesQuery } from "./pokemon/usePokemonWeaknessesQuery";
-export { usePokemonsQuery } from "./pokemons/usePokemonsQuery";
-export { usePokemonsPageQuery } from "./pokemons/usePokemonsPageQuery";
-export { usePokemonsByTypeQuery } from "./pokemons/usePokemonsByTypeQuery";
